Extract "inject_value" method into a base class
It will be shared with the bosh deployment manifest transmogrifier.
